refactor(ui): directory
change the directory `ShowSeasons`
